7311: Move PeerTaskFeatureToggle to more appropriate location
