+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/peertask/PeerTaskFeatureToggle.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.cli.custom;
+package org.hyperledger.besu.ethereum.eth.manager.peertask;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
